MOVE: types to own directory
