+++ src/hooks/useGetProducts.hook.ts
@@ -1,21 +1,6 @@
 import useSWR from 'swr';
 import { getCollection } from '../helpers';
-import { Timestamp } from 'firebase/firestore';
-
-export type Products = {
-  id: string;
-  name: string;
-  otherNames: string[];
-  category: string;
-  about: string;
-  images: string[];
-  price: number;
-  stock: number;
-  rating: 1 | 2 | 3 | 4 | 5;
-  createdAt: Timestamp;
-  updatedAt: Timestamp;
-  lastOrder: Timestamp;
-}[];
+import { Products } from '../types';
 
 /**
  * Retrieves all products from the database and returns an object with the
