+++ src/types/users.ts
@@ -1,0 +1,12 @@
+import { Timestamp } from 'firebase/firestore';
+
+export type Users = {
+  id: string;
+  name: string;
+  email: string;
+  address: string;
+  firstOrder: Timestamp;
+  lastOrder: Timestamp;
+  orders: number;
+  revenue: number;
+}[];
